Huge update - UUID & Stuff
Changed a lot of thinks so it should almost work with the new UUID
system. Some things still need to be fixed, but right now a lot changed.

There is also a new Spellmaker item: GetPlayerByUUID.

The UUID converter is added but may not work.

Soon there are coming more new things.

Keep hoping :)