Changed CreatureHp internals
Now the constructor wants also the maximum health of that entity to set
it.

There is added a new method for getting a creatures hp. You can wether
only pass the entity or also a maximum health. As the method creates a
new CreatureHp if there was no one found, that is necessary for
MythicMobs support.