Added new SpellCommand: GetProjectileName:CustomName
This commit adds a new feature to Ancient. Now projectiles that get
shoot can have custom names. The name comes from the ItemStack which
‚spawned‘ the projectile.
If you have a stack snowballs called ‚Test‘ and throw one you can get
the name of the snowball shoot when it hits something.

That gives you the opportunity to create different types of
projectiles, for example different types of grenades.