Fixed and added multiple things relating guilds
The command /guild setmotd is now in an other class.

Also there is a new command to remove the motd: /guild removemotd

/guild motd now says ‚Your guild has no motd‘ if the motd was removed
or not set

In AncientGuild the new commands were added (/guild removetag/motd).
Now tags also don’t have <> as brackets around it. You have to set
these on your own. Also a motd is only shown on join if there is one.

Last thing: A bug got fixed. In some cases Ancient stored guilds in the
folder plugins/Ancient and not in plugins/Ancient/Guilds. That is fixed
now. It is recommended to move all .guild files from plugins/Ancient
inside the folder plugins/Ancient/Guilds and remove the leading Guilds
of the file name.